+++ server/lib/models/tally/tally-model.wip.ts
@@ -1,16 +1,20 @@
 import { getDbClient } from 'server/lib/db.ts';
-import { type Tally } from 'generated/prisma/client';
 
 import { type RequestContext } from '../../request-context.ts';
 import { buildChangeRecord, logAuditEvent } from '../../audit-events.ts';
 import { AUDIT_EVENT_TYPE } from '../audit-event/consts.ts';
 
 import type { User } from '../user/user-model.ts';
+import type { Tally, PlainTally } from './types.ts';
 import { TALLY_STATE, type TallyMeasure } from './consts.ts';
+import { db2tally } from './helpers.ts';
 
 import { traced } from '../../metrics/tracer.ts';
 
-export type { Tally };
+export type {
+  Tally, PlainTally,
+};
+
 export type TallyData = {
   date: string;
   measure: TallyMeasure;
@@ -33,7 +37,7 @@ export class TallyModel {
   @traced
   static async getTallies(owner: User, filter: TallyFilter = {}): Promise<Tally[]> {
     const db = getDbClient();
-    const tallies = await db.tally.findMany({
+    const found = await db.tally.findMany({
       where: {
         ownerId: owner.id,
         state: TALLY_STATE.ACTIVE,
@@ -46,21 +50,29 @@ export class TallyModel {
         workId: (filter.workIds ?? []).length > 0 ? { in: filter.workIds } : undefined,
         tags: filter.tagIds ? { some: { id: { in: filter.tagIds } } } : undefined,
       },
+      include: {
+        tags: { select: { id: true } },
+      },
     });
 
+    const tallies = found.map(db2tally) as Tally[];
     return tallies;
   }
 
   @traced
   static async getTally(owner: User, id: number): Promise<Tally | null> {
     const db = getDbClient();
-    const tally = await db.tally.findUnique({
+    const found = await db.tally.findUnique({
       where: {
         id: id,
         ownerId: owner.id,
       },
+      include: {
+        tags: { select: { id: true } },
+      },
     });
 
+    const tally = db2tally(found);
     return tally;
   }
 
@@ -72,7 +84,7 @@ export class TallyModel {
 
   // TODO: move this to a batch creation domain
   @traced
-  static async batchCreateTallies(owner: User, data: BatchTallyData[], reqCtx: RequestContext): Promise<Tally[]> {
+  static async batchCreateTallies(owner: User, data: BatchTallyData[], reqCtx: RequestContext): Promise<PlainTally[]> {
     const db = getDbClient();
     const createds = await db.tally.createManyAndReturn({
       data: data.map(tallyData => ({
@@ -109,13 +121,17 @@ export class TallyModel {
   static async deleteTally(owner: User, tally: Tally, reqCtx: RequestContext): Promise<Tally> {
     // we actually delete deleted tallies
     const db = getDbClient();
-    const deleted = await db.tally.delete({
+    const dbDeleted = await db.tally.delete({
       where: {
         id: tally.id,
         ownerId: owner.id,
         state: TALLY_STATE.ACTIVE,
       },
+      include: {
+        tags: { select: { id: true } },
+      },
     });
+    const deleted = db2tally(dbDeleted)!;
 
     const changes = buildChangeRecord(tally, deleted);
     await logAuditEvent(
@@ -124,6 +140,6 @@ export class TallyModel {
       changes, reqCtx.sessionId,
     );
 
-    return deleted;
+    return tally;
   }
 }